fix #59, add #57
- fix widget disabling on os other than windows
- changed the emoji font to using `NotoEmoji` and added the font to assets
- added try catch to the creation of tray
- added option to not enable tray app through command line argument (`-`--no-tray`

+++ speech_translate/ui/frame/setting/general.py
@@ -488,7 +488,8 @@ class SettingGeneral:
         for file in os.listdir(dir_log):
             if file.endswith(".log"):
                 try:
-                    os.remove(os.path.join(dir_log, file))
+                    if file != current_log:
+                        os.remove(os.path.join(dir_log, file))
                 except Exception as e:
                     if file != current_log:  # show warning only if the fail to delete is not the current log
                         logger.warning("Failed to delete log file: " + file)
